requirements.txt
streamlit
openai>=1.0.0
PyPDF2
langchain-text-splitters
python-dotenv
